clip-service/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
ftfy==6.1.1
regex==2023.6.3
tqdm==4.65.0
Pillow==10.0.0
numpy==1.24.3
fastapi==0.99.1
uvicorn==0.22.0
python-multipart==0.0.6
chromadb==0.4.5
requests==2.31.0
aiohttp==3.8.4